requirements.txt
# Core numerics and data
numpy>=1.26,<3
pandas>=2.2,<3
scipy>=1.11,<2

# Classical ML
scikit-learn>=1.4,<2
joblib>=1.3,<2

# Config / YAML
PyYAML>=6.0.1,<7

# Plotting
matplotlib>=3.8,<4

# Hugging Face stack
transformers>=4.44,<5
tokenizers>=0.15,<0.20
datasets>=2.20,<3
accelerate>=0.33,<1

# Sentence embeddings
sentence-transformers>=2.7,<3

# Word vectors
gensim>=4.3,<5

# Deep learning
--index-url https://download.pytorch.org/whl/cu118
torch==2.2.2+cu118
torchvision==0.17.2+cu118
torchaudio==2.2.2+cu118

# Utilities
tqdm>=4.66,<5 
typing_extensions>=4.10,<5 
